Resolve Callback<->Model cycle via new BaseCallback
