+++ src/base_callbacks_test.ts
@@ -15,9 +15,10 @@
 // tslint:disable:max-line-length
 import {scalar} from '@tensorflow/tfjs-core';
 
-import {BaseLogger, CallbackList} from './callbacks';
+import {BaseLogger, CallbackList, History} from './base_callbacks';
+import {Callback} from './callbacks';
 import {disposeTensorsInLogs, resolveScalarsInLogs, UnresolvedLogs} from './engine/logs';
-import {History} from './engine/training';
+import {Model} from './engine/training';
 import {describeMathCPUAndGPU} from './utils/test_utils';
 
 // tslint:enable:max-line-length
@@ -174,5 +175,48 @@ describeMathCPUAndGPU('disposeTensorsInLogs', () => {
     expect(logs['c']).toEqual(-3);
     // tslint:disable-next-line:no-any
     expect((logs['d'] as any).isDisposed).toEqual(true);
+  });
+});
+
+
+describe('History Callback', () => {
+  it('onTrainBegin', async done => {
+    const history = new History();
+    await history.onTrainBegin();
+    expect(history.epoch).toEqual([]);
+    expect(history.history).toEqual({});
+    done();
+  });
+  it('onEpochEnd', async done => {
+    const history = new History();
+    await history.onTrainBegin();
+    await history.onEpochEnd(0, {'val_loss': 10, 'val_accuracy': 0.1});
+    expect(history.epoch).toEqual([0]);
+    expect(history.history).toEqual({'val_loss': [10], 'val_accuracy': [0.1]});
+    await history.onEpochEnd(1, {'val_loss': 9.5, 'val_accuracy': 0.2});
+    expect(history.epoch).toEqual([0, 1]);
+    expect(history.history)
+        .toEqual({'val_loss': [10, 9.5], 'val_accuracy': [0.1, 0.2]});
+    done();
+  });
+});
+
+class MockModel extends Model {
+  constructor(name: string) {
+    super({inputs: [], outputs: [], name});
+  }
+}
+
+class MockCallback extends Callback {}
+
+describe('CallbackList', () => {
+  it('Constructor and setModel with array of callbacks', () => {
+    const mockCallback1 = new MockCallback();
+    const mockCallback2 = new MockCallback();
+    const callbackList = new CallbackList([mockCallback1, mockCallback2]);
+    const model = new MockModel('MockModelA');
+    callbackList.setModel(model);
+    expect(mockCallback1.model).toEqual(model);
+    expect(mockCallback2.model).toEqual(model);
   });
 });
